pyproject.toml
[project]
name = "specloop"
version = "0.1.0"
description = "SpecLoop: Autonomous SpecKit Development Loop"
requires-python = ">=3.11"
dependencies = [
    "mcp>=0.1.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
    "click>=8.0.0",
]

[project.scripts]
specloop = "specloop.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/specloop"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = []

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
